EU-115 simplified code, moved business logic from users.services to common.utils
